Modify the slideshow slide change
The current slideshow fades one out while fading another in, which causes the div size to be 2X the height (or width) of the slide. The example fix for this is to set the slide positions to absolute 0 0 so they will lay on top of each other. Slide sizes, however, then have to be fixed instead of width=100% and height=auto. By hiding the existing slides in the container immediately then fading in the new slide (200 milliseconds), the effect is very similar, but now the slide container and slide sizes do not have to be fixed.